+++ internal/generic/thrift/parse.go
