+++ pkg/generic/thriftidl_provider_test.go
@@ -20,8 +20,8 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/cloudwego/kitex/internal/generic/thrift"
 	"github.com/cloudwego/kitex/internal/test"
-	"github.com/cloudwego/kitex/pkg/generic/thrift"
 )
 
 var testServiceContent = `
